Created empty auto group to replace example command as default autonomous mode command
